Start moving generic helpers to a helper file
