+++ scripts/generators/beats.py
@@ -1,4 +1,6 @@
+import ecs_helpers
 import yaml
+
 from collections import OrderedDict
 
 
@@ -13,7 +15,7 @@ def generate(ecs_nested, ecs_version):
             continue
         fieldset = ecs_nested[fieldset_name]
 
-        beats_field = dict_copy_keys_ordered(fieldset, allowed_fieldset_keys)
+        beats_field = ecs_helpers.dict_copy_keys_ordered(fieldset, allowed_fieldset_keys)
         beats_field['fields'] = fieldset_field_array(fieldset['fields'])
         beats_fields.append(beats_field)
 
@@ -33,7 +35,7 @@ def fieldset_field_array(source_fields):
     fields = []
     for nested_field_name in source_fields:
         ecs_field = source_fields[nested_field_name]
-        beats_field = dict_copy_keys_ordered(ecs_field, allowed_keys)
+        beats_field = ecs_helpers.dict_copy_keys_ordered(ecs_field, allowed_keys)
         beats_field['name'] = nested_field_name
         fields.append(beats_field)
     return sorted(fields, lambda x, y: cmp(x['name'], y['name']))
@@ -42,37 +44,9 @@ def fieldset_field_array(source_fields):
 
 
 def write_beats_yaml(beats_file, ecs_version):
+    warning = file_header().format(version=ecs_version)
+    ecs_helpers.yaml_dump('generated/beats/fields.ecs.yml', [beats_file], preamble=warning)
 
-    with open('generated/beats/fields.ecs.yml', 'w') as outfile:
-        outfile.write(file_header().format(version=ecs_version))
-        yaml.dump([beats_file], outfile, default_flow_style=False)
-
-
-def dict_copy_keys_ordered(dict, copied_keys):
-    ordered_dict = OrderedDict()
-    for key in copied_keys:
-        if key in dict:
-            ordered_dict[key] = dict[key]
-    return ordered_dict
-
-
-def indent(text, indent, char=' '):
-    padding = indent * char
-    return ''.join(padding + line for line in text.splitlines(True))
-
-# Rendering
-
-
-def yaml_ordereddict(dumper, data):
-    value = []
-    for item_key, item_value in data.items():
-        node_key = dumper.represent_data(item_key)
-        node_value = dumper.represent_data(item_value)
-        value.append((node_key, node_value))
-    return yaml.nodes.MappingNode(u'tag:yaml.org,2002:map', value)
-
-
-yaml.add_representer(OrderedDict, yaml_ordereddict)
 
 # Templates
